Rename checkout fragments and restructure Cart operations
