+++ packages/venia-ui/lib/components/CartPage/PriceAdjustments/ShippingMethods/shippingRadios.js
@@ -8,7 +8,7 @@ import RadioGroup from '../../../RadioGroup';
 import { CartPageFragment } from '../../cartPageFragments.gql';
 import ShippingRadio from './shippingRadio';
 import defaultClasses from './shippingRadios.css';
-import { SelectedShippingMethodFragment } from './shippingMethodsFragments';
+import { SelectedShippingMethodFragment } from './shippingMethodsFragments.gql';
 
 const ShippingRadios = props => {
     const {
